VRF-669: adding CTF tests for VRF v2.5 cancel subscription, oracle wi… (#11159)
* VRF-669: adding CTF tests for VRF v2.5 cancel subscription, oracle withdraw; added fund return as a teardown for the WASP load test

* VRF-669: updating tests to have better balance calculation

* VRF-669: small update

* VRF-669: enabling tests; making use of big.Int Cmp() method

* VRF-669: adding sub cancellation test for the coordinator owner; code refactoring; fixing load test

+++ integration-tests/load/vrfv2plus/config.go
@@ -57,8 +57,8 @@ type Funding struct {
 }
 
 type SubFunding struct {
-	SubFundsLink   int64 `toml:"sub_funds_link"`
-	SubFundsNative int64 `toml:"sub_funds_native"`
+	SubFundsLink   float64 `toml:"sub_funds_link"`
+	SubFundsNative float64 `toml:"sub_funds_native"`
 }
 
 type Soak struct {
